feat: add agent yaml to config from explorer
