+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 - Added Grok 4 model with token limit adjustments.
 - Extended support for Grok via dedicated handler and SDK updates.
 
+## 0.32.2 - 2025-07-12
+
+- Added "Add Agent to Config" context menu option in the TeX Agents view.
+
 ## 0.32.0 - 2025-07-07
 
 - Default model fix
