+++ src/explorer/ExplorerOperations.ts
@@ -12,6 +12,7 @@ import { agentDirectories } from '@frontend/agents/AgentDirectoryManager';
 import { AbsoluteFS } from '@utils/files';
 import { getConfig } from '@utils/config';
 import { FileItem } from './FileItem';
+import { validateYamlAndPromptAdd } from '@frontend/agents/register';
 
 const NEW_AGENT_TEMPLATE = `# --- Agent Inheritance (Optional) ---
 # inherits: base
@@ -284,6 +285,29 @@ export class ExplorerOperations {
           err,
         );
       }
+    }
+  }
+
+  async addToConfig(item: FileItem) {
+    if (!item) {
+      return;
+    }
+
+    if (path.extname(item.resourceUri.fsPath) !== '.yaml') {
+      vscode.window.showInformationMessage(
+        'Only YAML files can be added as agents.',
+      );
+      return;
+    }
+
+    try {
+      await validateYamlAndPromptAdd(item.resourceUri.fsPath, true);
+    } catch (err) {
+      await showLoggedErrorMessage(
+        CHANNEL,
+        'Failed to add agent to configuration',
+        err,
+      );
     }
   }
 }
